ActiveX Code Shuffling
Move the Control ActiveX interfaces to their own file and move the remaining interfaces to CsWin32.

Pull out the nested AxContainer.ExtenderProxy class.

Remove unused Interop code.

Some sporadic cleanup in code I've been reviewing.
