--- src/System.Windows.Forms.Primitives/src/Interop/Ole32/Interop.IPersist.cs
@@ -1,21 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Runtime.InteropServices;
-
-internal partial class Interop
-{
-    internal static partial class Ole32
-    {
-        [ComImport]
-        [Guid("0000010C-0000-0000-C000-000000000046")]
-        [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-        public unsafe interface IPersist
-        {
-            [PreserveSig]
-            HRESULT GetClassID(
-                Guid* pClassID);
-        }
-    }
-}
