--- src/System.Windows.Forms.Primitives/src/Interop/Ole32/Interop.OLEGETMONIKER.cs
@@ -1,17 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-internal static partial class Interop
-{
-    internal static partial class Ole32
-    {
-        public enum OLEGETMONIKER : uint
-        {
-            ONLYIFTHERE = 1,
-            FORCEASSIGN = 2,
-            UNASSIGN = 3,
-            TEMPFORUSER = 4
-        }
-    }
-}
