--- src/System.Windows.Forms.Primitives/src/Interop/Ole32/Interop.USERCLASSTYPE.cs
@@ -1,16 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-internal static partial class Interop
-{
-    internal static partial class Ole32
-    {
-        public enum USERCLASSTYPE : uint
-        {
-            FULL = 1,
-            SHORT = 2,
-            APPNAME = 3,
-        }
-    }
-}
