+++ src/System.Windows.Forms.Primitives/src/Windows/Win32/System/Com/StructuredStorage/IPropertyBag.cs
@@ -2,14 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-internal static partial class Interop
+namespace Windows.Win32.System.Com.StructuredStorage;
+
+internal unsafe partial struct IPropertyBag : IVTable<IPropertyBag, IPropertyBag.Vtbl>
 {
-    internal static partial class Ole32
-    {
-        public unsafe struct CA
-        {
-            public uint cElems;
-            public void* pElems;
-        }
-    }
 }
