+++ src/System.Windows.Forms/src/System/Windows/Forms/Control.cs
@@ -15,10 +15,8 @@ using System.Windows.Forms.Automation;
 using System.Windows.Forms.Layout;
 using System.Windows.Forms.Primitives;
 using Microsoft.Win32;
-using Windows.Win32.System.Com.StructuredStorage;
 using Windows.Win32.System.Ole;
 using static Interop;
-using Com = Windows.Win32.System.Com;
 using ComTypes = System.Runtime.InteropServices.ComTypes;
 using Encoding = System.Text.Encoding;
 
@@ -42,18 +40,6 @@ namespace System.Windows.Forms
     [ToolboxItemFilter("System.Windows.Forms")]
     public unsafe partial class Control :
         Component,
-        IOleControl.Interface,
-        IOleObject.Interface,
-        IOleInPlaceObject.Interface,
-        IOleInPlaceActiveObject.Interface,
-        IOleWindow.Interface,
-        Ole32.IViewObject,
-        Ole32.IViewObject2,
-        Com.IPersist.Interface,
-        Com.IPersistStreamInit.Interface,
-        IPersistPropertyBag.Interface,
-        IPersistStorage.Interface,
-        Ole32.IQuickActivate,
         ISupportOleDropSource,
         IDropTarget,
         ISynchronizeInvoke,
@@ -640,45 +626,6 @@ namespace System.Windows.Forms
                 //valid values are -1 to 0x40
                 SourceGenerated.EnumValidator.Validate(value);
                 Properties.SetInteger(s_accessibleRoleProperty, (int)value);
-            }
-        }
-
-        // Helper methods for retrieving an ActiveX properties.
-        // We abstract these to another method so we do not force JIT the ActiveX codebase.
-
-        private Color ActiveXAmbientBackColor => ActiveXInstance.AmbientBackColor;
-        private Color ActiveXAmbientForeColor => ActiveXInstance.AmbientForeColor;
-        private Font? ActiveXAmbientFont => ActiveXInstance.AmbientFont;
-        private bool ActiveXEventsFrozen => ActiveXInstance.EventsFrozen;
-        private IntPtr ActiveXHWNDParent => ActiveXInstance.HWNDParent;
-
-        /// <summary>
-        ///  Retrieves the ActiveX control implementation for this control.
-        ///  This will demand create the implementation if it does not already exist.
-        /// </summary>
-        private ActiveXImpl ActiveXInstance
-        {
-            get
-            {
-                ActiveXImpl? activeXImpl = (ActiveXImpl?)Properties.GetObject(s_activeXImplProperty);
-                if (activeXImpl is null)
-                {
-                    // Don't allow top level objects to be hosted
-                    // as activeX controls.
-                    if (GetState(States.TopLevel))
-                    {
-                        throw new NotSupportedException(SR.AXTopLevelSource);
-                    }
-
-                    activeXImpl = new ActiveXImpl(this);
-
-                    // PERF: IsActiveX is called quite a bit - checked everywhere from sizing to event raising.  Using a state
-                    // bit to track PropActiveXImpl instead of fetching from the property store.
-                    SetExtendedState(ExtendedStates.IsActiveX, true);
-                    Properties.SetObject(s_activeXImplProperty, activeXImpl);
-                }
-
-                return activeXImpl;
             }
         }
 
@@ -4558,42 +4505,6 @@ namespace System.Windows.Forms
             {
                 User32.NotifyWinEvent((uint)accEvent, new HandleRef(this, Handle), objectID, childID + 1);
             }
-        }
-
-        /// <summary>
-        ///  Helper method for retrieving an ActiveX property.  We abstract these
-        ///  to another method so we do not force JIT the ActiveX codebase.
-        /// </summary>
-        private Region ActiveXMergeRegion(Region region)
-        {
-            return ActiveXInstance.MergeRegion(region);
-        }
-
-        /// <summary>
-        ///  Helper method for retrieving an ActiveX property.  We abstract these
-        ///  to another method so we do not force JIT the ActiveX codebase.
-        /// </summary>
-        private void ActiveXOnFocus(bool focus)
-        {
-            ActiveXInstance.OnFocus(focus);
-        }
-
-        /// <summary>
-        ///  Helper method for retrieving an ActiveX property.  We abstract these
-        ///  to another method so we do not force JIT the ActiveX codebase.
-        /// </summary>
-        private void ActiveXViewChanged()
-        {
-            ActiveXInstance.ViewChangedInternal();
-        }
-
-        /// <summary>
-        ///  Helper method for retrieving an ActiveX property.  We abstract these
-        ///  to another method so we do not force JIT the ActiveX codebase.
-        /// </summary>
-        private void ActiveXUpdateBounds(ref int x, ref int y, ref int width, ref int height, SET_WINDOW_POS_FLAGS flags)
-        {
-            ActiveXInstance.UpdateBounds(ref x, ref y, ref width, ref height, flags);
         }
 
         /// <summary>
@@ -13575,783 +13486,6 @@ namespace System.Windows.Forms
         void ISupportOleDropSource.OnQueryContinueDrag(QueryContinueDragEventArgs queryContinueDragEventArgs)
         {
             OnQueryContinueDrag(queryContinueDragEventArgs);
-        }
-
-        unsafe HRESULT IOleControl.Interface.GetControlInfo(CONTROLINFO* pCI)
-        {
-            if (pCI is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetControlInfo");
-            pCI->cb = (uint)sizeof(CONTROLINFO);
-            pCI->hAccel = HACCEL.Null;
-            pCI->cAccel = 0;
-            pCI->dwFlags = 0;
-
-            if (IsInputKey(Keys.Return))
-            {
-                pCI->dwFlags |= CTRLINFO.CTRLINFO_EATS_RETURN;
-            }
-
-            if (IsInputKey(Keys.Escape))
-            {
-                pCI->dwFlags |= CTRLINFO.CTRLINFO_EATS_ESCAPE;
-            }
-
-            return ActiveXInstance.GetControlInfo(pCI);
-        }
-
-        unsafe HRESULT IOleControl.Interface.OnMnemonic(MSG* pMsg)
-        {
-            if (pMsg is null)
-            {
-                return HRESULT.E_INVALIDARG;
-            }
-
-            // If we got a mnemonic here, then the appropriate control will focus itself which
-            // will cause us to become UI active.
-            bool processed = ProcessMnemonic((char)(nuint)pMsg->wParam);
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:OnMnemonic processed: {processed}");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleControl.Interface.OnAmbientPropertyChange(int dispID)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:OnAmbientPropertyChange. Dispid: {dispID}");
-            Debug.Indent();
-            ActiveXInstance.OnAmbientPropertyChange((Ole32.DispatchID)dispID);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleControl.Interface.FreezeEvents(BOOL bFreeze)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:FreezeEvents. Freeze: {bFreeze}");
-            ActiveXInstance.EventsFrozen = bFreeze;
-            Debug.Assert(ActiveXInstance.EventsFrozen == bFreeze, "Failed to set EventsFrozen correctly");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleInPlaceActiveObject.Interface.GetWindow(HWND* phwnd)
-        {
-            return ((IOleInPlaceObject.Interface)this).GetWindow(phwnd);
-        }
-
-        HRESULT IOleInPlaceActiveObject.Interface.ContextSensitiveHelp(BOOL fEnterMode)
-        {
-            return ((IOleInPlaceObject.Interface)this).ContextSensitiveHelp(fEnterMode);
-        }
-
-        unsafe HRESULT IOleInPlaceActiveObject.Interface.TranslateAccelerator(MSG* lpmsg)
-        {
-            return ActiveXInstance.TranslateAccelerator(lpmsg);
-        }
-
-        HRESULT IOleInPlaceActiveObject.Interface.OnFrameWindowActivate(BOOL fActivate)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:OnFrameWindowActivate");
-            OnFrameWindowActivate(fActivate);
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleInPlaceActiveObject.Interface.OnDocWindowActivate(BOOL fActivate)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:OnDocWindowActivate.  Activate: {(bool)fActivate}");
-            Debug.Indent();
-            ActiveXInstance.OnDocWindowActivate(fActivate);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleInPlaceActiveObject.Interface.ResizeBorder(RECT* prcBorder, IOleInPlaceUIWindow* pUIWindow, BOOL fFrameWindow)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:ResizesBorder");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleInPlaceActiveObject.Interface.EnableModeless(BOOL fEnable)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:EnableModeless");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT IOleInPlaceObject.Interface.GetWindow(HWND* phwnd)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetWindow");
-            HRESULT hr = ActiveXInstance.GetWindow(phwnd);
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"\twin == {(phwnd is null ? HWND.Null : *phwnd)}");
-            return hr;
-        }
-
-        HRESULT IOleInPlaceObject.Interface.ContextSensitiveHelp(BOOL fEnterMode)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:ContextSensitiveHelp.  Mode: {fEnterMode}");
-            if (fEnterMode)
-            {
-                OnHelpRequested(new HelpEventArgs(MousePosition));
-            }
-
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleInPlaceObject.Interface.InPlaceDeactivate()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:InPlaceDeactivate");
-            Debug.Indent();
-            HRESULT hr = ActiveXInstance.InPlaceDeactivate();
-            Debug.Unindent();
-            return hr;
-        }
-
-        HRESULT IOleInPlaceObject.Interface.UIDeactivate()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:UIDeactivate");
-            return ActiveXInstance.UIDeactivate();
-        }
-
-        unsafe HRESULT IOleInPlaceObject.Interface.SetObjectRects(RECT* lprcPosRect, RECT* lprcClipRect)
-        {
-            if (lprcClipRect is not null)
-            {
-                Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo,
-                    $"AxSource:SetObjectRects({lprcClipRect->left}, {lprcClipRect->top}, {lprcClipRect->right}, {lprcClipRect->bottom})");
-            }
-
-            Debug.Indent();
-            HRESULT hr = ActiveXInstance.SetObjectRects(lprcPosRect, lprcClipRect);
-            Debug.Unindent();
-            return hr;
-        }
-
-        HRESULT IOleInPlaceObject.Interface.ReactivateAndUndo()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:ReactivateAndUndo");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.SetClientSite(IOleClientSite* pClientSite)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetClientSite");
-            ActiveXInstance.SetClientSite(pClientSite);
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetClientSite(IOleClientSite** ppClientSite)
-        {
-            if (ppClientSite is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetClientSite");
-            var clientSite = ActiveXInstance.GetClientSite();
-            *ppClientSite = null;
-            if (clientSite is not null)
-            {
-                bool result = ComHelpers.TryGetComPointer(clientSite, out *ppClientSite);
-                Debug.Assert(result);
-            }
-
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleObject.Interface.SetHostNames(PCWSTR szContainerApp, PCWSTR szContainerObj)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetHostNames");
-
-            // Since ActiveX controls never "open" for editing, we shouldn't need to store these.
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleObject.Interface.Close(OLECLOSE dwSaveOption)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Close. Save option: " + dwSaveOption);
-            ActiveXInstance.Close(dwSaveOption);
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.SetMoniker(OLEWHICHMK dwWhichMoniker, Com.IMoniker* pmk)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetMoniker");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetMoniker(OLEGETMONIKER dwAssign, OLEWHICHMK dwWhichMoniker, Com.IMoniker** ppmk)
-        {
-            if (ppmk is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetMoniker");
-            *ppmk = null;
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT IOleObject.Interface.InitFromData(Com.IDataObject* pDataObject, BOOL fCreation, uint dwReserved)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:InitFromData");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetClipboardData(uint dwReserved, Com.IDataObject** ppDataObject)
-        {
-            if (ppDataObject is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetClipboardData");
-            ppDataObject = null;
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT IOleObject.Interface.DoVerb(
-            int iVerb,
-            MSG* lpmsg,
-            IOleClientSite* pActiveSite,
-            int lindex,
-            HWND hwndParent,
-            RECT* lprcPosRect)
-        {
-            // In Office they are internally casting an iVerb to a short and not doing the proper sign extension.
-            short sVerb = unchecked((short)iVerb);
-            iVerb = sVerb;
-
-#if DEBUG
-            if (CompModSwitches.ActiveX.TraceInfo)
-            {
-                Debug.WriteLine("AxSource:DoVerb {");
-                Debug.WriteLine($"     verb: {iVerb}");
-                Debug.WriteLine($"     msg: {*lpmsg}");
-                Debug.WriteLine($"     activeSite: {*pActiveSite}");
-                Debug.WriteLine($"     index: {lindex}");
-                Debug.WriteLine($"     hwndParent: {hwndParent}");
-                Debug.WriteLine($"     posRect: {(lprcPosRect is null ? "null" : lprcPosRect->ToString())}");
-            }
-#endif
-            Debug.Indent();
-            try
-            {
-                return ActiveXInstance.DoVerb((Ole32.OLEIVERB)iVerb, lpmsg, pActiveSite, lindex, hwndParent, lprcPosRect);
-            }
-            finally
-            {
-                Debug.Unindent();
-                Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "}");
-            }
-        }
-
-        unsafe HRESULT IOleObject.Interface.EnumVerbs(IEnumOLEVERB** ppEnumOleVerb)
-        {
-            if (ppEnumOleVerb is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:EnumVerbs");
-            HRESULT hr = ActiveXImpl.EnumVerbs(out Ole32.IEnumOLEVERB oleVerb);
-            bool result = ComHelpers.TryGetComPointer(oleVerb, out *ppEnumOleVerb);
-            Debug.Assert(result);
-            return hr;
-        }
-
-        HRESULT IOleObject.Interface.Update()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:OleUpdate");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleObject.Interface.IsUpToDate()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IsUpToDate");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetUserClassID(Guid* pClsid)
-        {
-            if (pClsid is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            *pClsid = GetType().GUID;
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetUserClassID.  ClassID: " + pClsid->ToString());
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetUserType(USERCLASSTYPE dwFormOfType, PWSTR* pszUserType)
-        {
-            if (pszUserType is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetUserType");
-            *pszUserType = (char*)Marshal.StringToCoTaskMemUni(
-                dwFormOfType == USERCLASSTYPE.USERCLASSTYPE_FULL ? GetType().FullName : GetType().Name);
-
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.SetExtent(Com.DVASPECT dwDrawAspect, SIZE* psizel)
-        {
-            if (psizel is null)
-            {
-                return HRESULT.E_INVALIDARG;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:SetExtent({psizel->Width}, {psizel->Height}");
-            Debug.Indent();
-            ActiveXInstance.SetExtent((Ole32.DVASPECT)dwDrawAspect, (Size*)psizel);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetExtent(Com.DVASPECT dwDrawAspect, SIZE* psizel)
-        {
-            if (psizel is null)
-            {
-                return HRESULT.E_INVALIDARG;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:GetExtent.  Aspect: {dwDrawAspect.ToString()}");
-            Debug.Indent();
-            ActiveXInstance.GetExtent((Ole32.DVASPECT)dwDrawAspect, (Size*)psizel);
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"value: {psizel->Width}, {psizel->Height}");
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.Advise(Com.IAdviseSink* pAdvSink, uint* pdwConnection)
-        {
-            if (pdwConnection is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Advise");
-            *pdwConnection = ActiveXInstance.Advise(pAdvSink);
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IOleObject.Interface.Unadvise(uint dwConnection)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Unadvise");
-            Debug.Indent();
-            HRESULT hr = ActiveXInstance.Unadvise(dwConnection);
-            Debug.Unindent();
-            return hr;
-        }
-
-        unsafe HRESULT IOleObject.Interface.EnumAdvise(Com.IEnumSTATDATA** ppenumAdvise)
-        {
-            if (ppenumAdvise is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:EnumAdvise");
-            *ppenumAdvise = null;
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT IOleObject.Interface.GetMiscStatus(Com.DVASPECT dwAspect, OLEMISC* pdwStatus)
-        {
-            if (pdwStatus is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            if (!dwAspect.HasFlag(Com.DVASPECT.DVASPECT_CONTENT))
-            {
-                Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetMiscStatus.  Status: ERROR, wrong aspect.");
-                *pdwStatus = 0;
-                return HRESULT.DV_E_DVASPECT;
-            }
-
-            OLEMISC status = OLEMISC.OLEMISC_ACTIVATEWHENVISIBLE | OLEMISC.OLEMISC_INSIDEOUT | OLEMISC.OLEMISC_SETCLIENTSITEFIRST;
-            if (GetStyle(ControlStyles.ResizeRedraw))
-            {
-                status |= OLEMISC.OLEMISC_RECOMPOSEONRESIZE;
-            }
-
-            if (this is IButtonControl)
-            {
-                status |= OLEMISC.OLEMISC_ACTSLIKEBUTTON;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:GetMiscStatus. Status: {status}");
-            *pdwStatus = status;
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleObject.Interface.SetColorScheme(LOGPALETTE* pLogpal)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetColorScheme");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IOleWindow.Interface.GetWindow(HWND* phwnd)
-        {
-            return ((IOleInPlaceObject.Interface)this).GetWindow(phwnd);
-        }
-
-        HRESULT IOleWindow.Interface.ContextSensitiveHelp(BOOL fEnterMode)
-        {
-            return ((IOleInPlaceObject.Interface)this).ContextSensitiveHelp(fEnterMode);
-        }
-
-        unsafe HRESULT Com.IPersist.Interface.GetClassID(Guid* pClassID)
-        {
-            if (pClassID is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            *pClassID = GetType().GUID;
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:IPersist.GetClassID.  ClassID: {*pClassID}");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IPersistPropertyBag.Interface.InitNew()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistPropertyBag.InitNew");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IPersistPropertyBag.Interface.GetClassID(Guid* pClassID)
-        {
-            if (pClassID is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            *pClassID = GetType().GUID;
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:IPersistPropertyBag.GetClassID.  ClassID: {*pClassID}");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IPersistPropertyBag.Interface.Load(IPropertyBag* pPropBag, Com.IErrorLog* pErrorLog)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Load (IPersistPropertyBag)");
-            Debug.Indent();
-            ActiveXInstance.Load((IPropertyBag.Interface)Marshal.GetObjectForIUnknown((nint)pPropBag), pErrorLog);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT IPersistPropertyBag.Interface.Save(IPropertyBag* pPropBag, BOOL fClearDirty, BOOL fSaveAllProperties)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Save (IPersistPropertyBag)");
-            Debug.Indent();
-            ActiveXInstance.Save((IPropertyBag.Interface)Marshal.GetObjectForIUnknown((nint)pPropBag), fClearDirty, fSaveAllProperties);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IPersistStorage.Interface.GetClassID(Guid* pClassID)
-        {
-            if (pClassID is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            *pClassID = GetType().GUID;
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:IPersistStorage.GetClassID.  ClassID: {*pClassID}");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IPersistStorage.Interface.IsDirty()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStorage.IsDirty");
-            return ActiveXInstance.IsDirty();
-        }
-
-        HRESULT IPersistStorage.Interface.InitNew(IStorage* pStg)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStorage.InitNew");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IPersistStorage.Interface.Load(IStorage* pStg)
-        {
-            if (pStg is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStorage.Load");
-            Debug.Indent();
-            HRESULT result = ActiveXInstance.Load(pStg);
-            Debug.Unindent();
-            return result;
-        }
-
-        HRESULT IPersistStorage.Interface.Save(IStorage* pStgSave, BOOL fSameAsLoad)
-        {
-            if (pStgSave is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStorage.Save");
-            Debug.Indent();
-            HRESULT result = ActiveXInstance.Save(pStgSave, fSameAsLoad);
-            Debug.Unindent();
-            return result;
-        }
-
-        HRESULT IPersistStorage.Interface.SaveCompleted(IStorage* pStgNew)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStorage.SaveCompleted");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT IPersistStorage.Interface.HandsOffStorage()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStorage.HandsOffStorage");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT Com.IPersistStreamInit.Interface.GetClassID(Guid* pClassID)
-        {
-            if (pClassID is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            *pClassID = GetType().GUID;
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, $"AxSource:IPersistStreamInit.GetClassID.  ClassID: {*pClassID}");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT Com.IPersistStreamInit.Interface.IsDirty()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStreamInit.IsDirty");
-            return ActiveXInstance.IsDirty();
-        }
-
-        HRESULT Com.IPersistStreamInit.Interface.Load(Com.IStream* pStm)
-        {
-            if (pStm is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStreamInit.Load");
-            Debug.Indent();
-            ActiveXInstance.Load(pStm);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        HRESULT Com.IPersistStreamInit.Interface.Save(Com.IStream* pStm, BOOL fClearDirty)
-        {
-            if (pStm is null)
-            {
-                return HRESULT.E_POINTER;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStreamInit.Save");
-            Debug.Indent();
-            ActiveXInstance.Save(pStm, fClearDirty);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        HRESULT Com.IPersistStreamInit.Interface.GetSizeMax(ulong* pCbSize)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetSizeMax");
-            return HRESULT.S_OK;
-        }
-
-        HRESULT Com.IPersistStreamInit.Interface.InitNew()
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:IPersistStreamInit.InitNew");
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT Ole32.IQuickActivate.QuickActivate(Ole32.QACONTAINER pQaContainer, Ole32.QACONTROL* pQaControl)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:QuickActivate");
-            Debug.Indent();
-            HRESULT hr = ActiveXInstance.QuickActivate(pQaContainer, pQaControl);
-            Debug.Unindent();
-            return hr;
-        }
-
-        unsafe HRESULT Ole32.IQuickActivate.SetContentExtent(Size* pSizel)
-        {
-            if (pSizel is null)
-            {
-                return HRESULT.E_INVALIDARG;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetContentExtent");
-            Debug.Indent();
-            ActiveXInstance.SetExtent(Ole32.DVASPECT.CONTENT, pSizel);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT Ole32.IQuickActivate.GetContentExtent(Size* pSizel)
-        {
-            if (pSizel is null)
-            {
-                return HRESULT.E_INVALIDARG;
-            }
-
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetContentExtent");
-            Debug.Indent();
-            ActiveXInstance.GetExtent(Ole32.DVASPECT.CONTENT, pSizel);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT Ole32.IViewObject.Draw(
-            Ole32.DVASPECT dwDrawAspect,
-            int lindex,
-            IntPtr pvAspect,
-            Ole32.DVTARGETDEVICE* ptd,
-            IntPtr hdcTargetDev,
-            IntPtr hdcDraw,
-            RECT* lprcBounds,
-            RECT* lprcWBounds,
-            IntPtr pfnContinue,
-            uint dwContinue)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Draw");
-
-            Debug.Indent();
-            HRESULT hr = ActiveXInstance.Draw(
-                dwDrawAspect,
-                lindex,
-                pvAspect,
-                ptd,
-                hdcTargetDev,
-                hdcDraw,
-                lprcBounds,
-                lprcWBounds,
-                pfnContinue,
-                dwContinue);
-            Debug.Unindent();
-            return HRESULT.S_OK;
-        }
-
-        unsafe HRESULT Ole32.IViewObject.GetColorSet(
-            Ole32.DVASPECT dwDrawAspect,
-            int lindex,
-            IntPtr pvAspect,
-            Ole32.DVTARGETDEVICE* ptd,
-            IntPtr hicTargetDev,
-            LOGPALETTE* ppColorSet)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetColorSet");
-
-            // GDI+ doesn't do palettes.
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT Ole32.IViewObject.Freeze(Ole32.DVASPECT dwDrawAspect, int lindex, IntPtr pvAspect, uint* pdwFreeze)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Freezes");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        HRESULT Ole32.IViewObject.Unfreeze(uint dwFreeze)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Unfreeze");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        HRESULT Ole32.IViewObject.SetAdvise(Ole32.DVASPECT aspects, Ole32.ADVF advf, ComTypes.IAdviseSink pAdvSink)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetAdvise");
-            return ActiveXInstance.SetAdvise(aspects, advf, pAdvSink);
-        }
-
-        unsafe HRESULT Ole32.IViewObject.GetAdvise(Ole32.DVASPECT* pAspects, Ole32.ADVF* pAdvf, ComTypes.IAdviseSink[] ppAdvSink)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetAdvise");
-            return ActiveXInstance.GetAdvise(pAspects, pAdvf, ppAdvSink);
-        }
-
-        unsafe HRESULT Ole32.IViewObject2.Draw(
-            Ole32.DVASPECT dwDrawAspect,
-            int lindex,
-            IntPtr pvAspect,
-            Ole32.DVTARGETDEVICE* ptd,
-            IntPtr hdcTargetDev,
-            IntPtr hdcDraw,
-            RECT* lprcBounds,
-            RECT* lprcWBounds,
-            IntPtr pfnContinue,
-            uint dwContinue)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Draw");
-            Debug.Indent();
-            HRESULT hr = ActiveXInstance.Draw(
-                dwDrawAspect,
-                lindex,
-                pvAspect,
-                ptd,
-                hdcTargetDev,
-                hdcDraw,
-                lprcBounds,
-                lprcWBounds,
-                pfnContinue,
-                dwContinue);
-            Debug.Unindent();
-            return hr;
-        }
-
-        unsafe HRESULT Ole32.IViewObject2.GetColorSet(
-            Ole32.DVASPECT dwDrawAspect,
-            int lindex,
-            IntPtr pvAspect,
-            Ole32.DVTARGETDEVICE* ptd,
-            IntPtr hicTargetDev,
-            LOGPALETTE* ppColorSet)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetColorSet");
-
-            // GDI+ doesn't do palettes.
-            return HRESULT.E_NOTIMPL;
-        }
-
-        unsafe HRESULT Ole32.IViewObject2.Freeze(Ole32.DVASPECT dwDrawAspect, int lindex, IntPtr pvAspect, uint* pdwFreeze)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Freezes");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        HRESULT Ole32.IViewObject2.Unfreeze(int dwFreeze)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:Unfreeze");
-            return HRESULT.E_NOTIMPL;
-        }
-
-        HRESULT Ole32.IViewObject2.SetAdvise(Ole32.DVASPECT aspects, Ole32.ADVF advf, ComTypes.IAdviseSink pAdvSink)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:SetAdvise");
-            return ActiveXInstance.SetAdvise(aspects, advf, pAdvSink);
-        }
-
-        unsafe HRESULT Ole32.IViewObject2.GetAdvise(Ole32.DVASPECT* pAspects, Ole32.ADVF* pAdvf, ComTypes.IAdviseSink[] ppAdvSink)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetAdvise");
-            return ActiveXInstance.GetAdvise(pAspects, pAdvf, ppAdvSink);
-        }
-
-        unsafe HRESULT Ole32.IViewObject2.GetExtent(Ole32.DVASPECT dwDrawAspect, int lindex, Ole32.DVTARGETDEVICE* ptd, Size* lpsizel)
-        {
-            Debug.WriteLineIf(CompModSwitches.ActiveX.TraceInfo, "AxSource:GetExtent (IViewObject2)");
-
-            // We already have an implementation of this [from IOleObject]
-            return ((IOleObject.Interface)this).GetExtent((Com.DVASPECT)dwDrawAspect, (SIZE*)lpsizel);
         }
 
         #region IKeyboardToolTip implementation
